Adjusts sums for annual expenses.
The sum of expenses listed in the summary didn't match the number listed in
the expense itemization. I am assuming that this is due to not enumerated
individual expense items prior to the asset transfer to SPI. Unfortunately,
we lost the ability to retrieve historical bank statements after we closed down
our account. The difference thus unaccounted as individual items is $125.66.