Second revision of OBF Bylaws, part 1.
This set of proposed revisions to the OBF Bylaws makes the following
changes:

* Remove the Parliamentarian officer, and reassigns (or removes, see
  below) their roles.
* Removes Nominating Committee, and simplifies nomination process.
* Simplifies process for Bylaws review by giving authority on how to
  conduct the review to the Board.

The revision maintains all first (Article) and second (Item) level
numbering so as not to possibly change a reference (external or internal)
to an Article or Item it was not originally meant to reference. Therefore,
rather than renumbering subsequent items, deleted items are maintained but
listed as _Removed_. This also helps to clarify the diff between versions.