Factors in $500 expense for awards.
The expense (under programmatic activities for BOSC) had been listed
in the report, and had been included in the balance accounting spreadsheet
provided by ISCB, but due an error in a subtotal formula wasn't included
in the calculation of the corresponding totals. This error was only
corrected in the fall of 2014.