Allows for other projects to be grandfathered
We have been approached in the past by a number of different projects
about how to become an OBF member. That there wasn't a policy when they
asked wasn't their fault, and hence arguably there should be some kind
of fast-track path for those. This change gives the Board broad powers
to grandfather pre-existing projects, but the most typical use of this
will be in essence "fast-track" projects that wanted to become members
in the past, by grandfathering them into Candidate OBF Project status
instead of them having to wait for the next community vote to attain
it, and then having to wait another year before they are eligible for
a vote on Affiliated status.