Affiliated projects policy (#44)
* Draft of affiliated projects policy

* Reorganization to prepare for merge into obf-docs repo.

* Changing title, background, and motivation.

* Ontologies project is not different from other projects, hence simplify.

* Don't categorize and state policy on non-voted categories. Simplify instead.

* Simplify voting preamble.

Maybe this needs more details, but that's not clear yet.

* Consolidate trailing items into Maintenance of status and Dispute resolution.

* Change voted-on categories to Core, Affiliated, and Candidate.

* Replacing O|B|F with OBF, according to Peter's comment.

* Small corrections in response to review comments.

Changed left over occurrence of 'main' to 'core'. Changed 'largisch' to 'established'. Both suggestions were from Peter.

* Further consistency corrections for terminology.

* Minor grammatical and stylistic fixes.

* Consolidate and streamline status descriptions and voting process.

* Adding the Affiliated Project policy document.

* Small polish in the document header.

* add some guidance as to benefits and requirements for affiliate status

* fix broken link for projects list

* add notes about inactive projects, some reformatting:

* change inactivity period to twelve menths and add FAQ section.

* Update Affiliated-project-FAQ.md

* add links to financial sponsors

* Update Affiliated-project-FAQ.md

* added obf to title; other minor wording tweaks

always helpful to give some context for people who find the file without going through the hierarchy, e.g. by googling (does google index github?)

anyway, thanks for starting this, looks great!

* Streamlines and clarifies motivations

* Moves benefits from policy to FAQ.

Also clarifies the financial management benefits FAQ (or more precisely,
the answer).

* Clarifies FAQ about signing over ownership

* Various small language and formatting tweaks

* Various language clarifications and fixes

* Create issue template for candidate project application

* fix instructions about how to apply for candidate membership status. Addresses #45

* Minor language tweaks for Affiliate Project policy and FAQ

* Removes Core OBF Project tier

As per #14 and #23 discussion threads, this removes the differentiation
between Core and Affiliated OBF Project, and retains only Candidate and
Affiliate statuses.

This includes additional changes in the voting and status maintenance
sections to make it all more coherent and consistent.

* Tweaks links and references so they are up to date

* Adds clause on grandfathered projects

Originally, the Core OBF Project tier was going to be seeded by
grandfathering the "original" OBF member projects. Now that this tier
has been removed (see #55), this section states that these projects
will be grandfathered to Affiliated OBF Project status.

* Tweaks application instructions for consistency  (#59)

Tweaks wording to match application template wording and single affiliated project
status. Removes redundant information

Also ensures that it is clear that a template must be used to apply for candidate status, not just creating an unstructured github issue.

* Allows for other projects to be grandfathered

We have been approached in the past by a number of different projects
about how to become an OBF member. That there wasn't a policy when they
asked wasn't their fault, and hence arguably there should be some kind
of fast-track path for those. This change gives the Board broad powers
to grandfather pre-existing projects, but the most typical use of this
will be in essence "fast-track" projects that wanted to become members
in the past, by grandfathering them into Candidate OBF Project status
instead of them having to wait for the next community vote to attain
it, and then having to wait another year before they are eligible for
a vote on Affiliated status.

* Adds link for Google Summer of Code

Thanks @heuermh for spotting and the correction.

* Updates status of policy document after Board and Membership approval

Co-authored-by: Pjotr Prins <pjotr.public01@thebird.nl>
Co-authored-by: Yo Yehudi <yochannah@gmail.com>
Co-authored-by: Nomi Harris <nlharris@users.noreply.github.com>